New help about options: Licence information Changed implementation of RvMsg string serialization to allow later implemetation of editing and resending messages.
Rvsn00p now uses the first found of:
- Tibco Rvscript tibrvXmlConvert
- MTree.toString()
- TibRvMsg.toString()

This means that if the rvscript.jar is in the classpath tibrvXmlConvert will be used as
string serializer. If it is not found the serializer will be downgraded to the Tibco SDK
AE Mtree format and if that is not not fount the standard RV API TibRvMsg will be used.