Added ConnectionList class to handle displaying the connection list. Added ImageFactory to handle loading and caching icons and banners.
Added RecordLedgerSelectionEvent and RecordLedgerSelectionListener, these are
going to form the basis of handling any actions which need to know about the
selection status of the ledger. A future release will do the same for the
connections list and the subject hierarchy.

Added NLSUtils and started to internationalize all of the strings. I plan to
update all of the actions for I18N as I move them to the new org.rvsnoop...
package tree, after that I'll go through the rest of the app and change over
any other strings. All images loaded via the new ImageFactory will be I18N
enabled automatically.

This is a work in progress! ButI've been away from the net for a few days
and want to get this checked in even iof it isn't complete.